==> 0-Extract five txt to xlsx format.ipynb <==
#### PubMed Cancer

#file = "../src/pubmed-Top1-lungcancer-set.txt"
InputDir = "../Src/"
Inputfile1 ="pubmed-cancer-set_10000.txt"
#Inputfile2 ="pubmed-COVID-19-set_10000.txt"
#Inputfile3 ="pubmed-genome-set_10000.txt"
#Inputfile4 ="pubmed-vaccine-set_10000.txt"
#Inputfile5 ="pubmed-virus-set_10000.txt"
#Inputfile = "pubmed-Top4_Prostateca-set.txt"

InputFileLocation = InputDir+Inputfile1
#InputFileLocation2 = InputDir+Inputfile2
#InputFileLocation3 = InputDir+Inputfile3
#InputFileLocation4 = InputDir+Inputfile4
#InputFileLocation5 = InputDir+Inputfile5
#InputFileLocation = InputFileLocation1+InputFileLocation2+InputFileLocation3+InputFileLocation4+InputFileLocation5
#with open (file, 'r',encoding="utf-8") as file_input:
with open (InputFileLocation, 'r',encoding="utf-8") as file_input:
    Lines = file_input.readlines()
    #print (Lines)
cnt = 0
LineIndex = 0
PMID_LineIndex = []
for OneLine in Lines:
    LineIndex = LineIndex + 1
    if (OneLine.startswith('PMID-')):
        cnt = cnt +1
        PMID_LineIndex.append(LineIndex)
PMID_LineIndex.append(LineIndex)
#print(cnt)    
#print(PMID_LineIndex)

CancerType = "Cancer"
Inputfile ="pubmed-cancer-set_10000.txt"
OutputFile = "pubmed-cancer-set_10000.xlsx"



InputDir = "../Src/"
OutputDir = "../Output/"

PUMID_List = []
TI_List = []
AB_List = []

#file = "../Src/pubmed-Top1-lungcancer-set.txt"
file = "../Src/pubmed-cancer-set_10000.txt"
with open (file, 'r',encoding="utf-8") as file_input:
    for i in range(0,cnt):
        StartLineIndex = PMID_LineIndex[i]
        EndLineIndex = PMID_LineIndex[i+1]
        OnePUMID = []
        OneTI_Info = []
        OneAB_Info = []
        for j in range(StartLineIndex,EndLineIndex):
            OneLine = file_input.readline()
            #OneLine = OneLine.strip()
            #print(OneLine)
            OnePUMID.append(OneLine)
        #print(len(OnePUMID))
        OnePubMed_LineCnt = len(OnePUMID)
        #print(OnePUMID)
        #print(OnePubMed_LineCnt)
        #OnePUMID_Info = []
        k = 0
        while (k < OnePubMed_LineCnt):
            if (OnePUMID[k].startswith('PMID-')):
                OnePUMID_Info = OnePUMID[k].split('- ')
                #print(str(k)+' PMID:'+OnePUMID_Info[1])
            if (OnePUMID[k].startswith('TI  -') or OnePUMID[k].startswith('Ti  -')):
                OneTI_Info = OnePUMID[k].split('- ')
                OneTI_Info[1] = OneTI_Info[1].strip()
                while (OnePUMID[k+1].startswith('  ')):
                    Temp = OnePUMID[k+1].split('   ')
                    #print(OnePUMID[k+1])
                    OneTI_Info[1] = OneTI_Info[1]+' '+Temp[1].strip()
                    k += 1
                #print(str(k)+' TI:'+OneTI_Info[1])
            if (OnePUMID[k].startswith('AB  -')):
                OneAB_Info = OnePUMID[k].split('- ')
                OneAB_Info[1] = OneAB_Info[1].strip()
                while (OnePUMID[k+1].startswith('  ')):
                    Temp = OnePUMID[k+1].split('   ')
                    #print(OnePUMID[k+1])
                    OneAB_Info[1] = OneAB_Info[1]+' '+Temp[1].strip()
                    k += 1
                #print(str(k)+' AB:'+OneTI_Info[1])
            k += 1

        PUMID_List.append(OnePUMID_Info[1].strip())
        TI_List.append(OneTI_Info[1].strip())
        if (len(OneAB_Info) == 2):
            AB_List.append(OneAB_Info[1].strip())
        else:
            AB_List.append('')
            #print(OnePUMID_Info[1]+"No Abstract")
        #AB_List.append(OneAB_Info[1].strip())
#PUMID_List
#TI_List
#AB_List
PubMed_Data = [[PUMID_List[i],TI_List[i],AB_List[i]] for i in range(cnt)]
#PubMed_Data

import pandas as pd

PubMed = pd.DataFrame(PubMed_Data, columns=['PUMID','TI','AB'])
#PubMed

PubMed['Type'] = CancerType

PubMed.to_excel("pubmed-cancer-set_10000.xlsx")

#### PubMed COVID-19

#file = "../src/pubmed-Top1-lungcancer-set.txt"
InputDir = "../Src/"
#Inputfile1 ="pubmed-cancer-set_10000.txt"
Inputfile2 ="pubmed-COVID-19-set_10000.txt"
#Inputfile3 ="pubmed-genome-set_10000.txt"
#Inputfile4 ="pubmed-vaccine-set_10000.txt"
#Inputfile5 ="pubmed-virus-set_10000.txt"
#Inputfile = "pubmed-Top4_Prostateca-set.txt"

#InputFileLocation = InputDir+Inputfile1
InputFileLocation = InputDir+Inputfile2
#InputFileLocation = InputDir+Inputfile3
#InputFileLocation = InputDir+Inputfile4
#InputFileLocation = InputDir+Inputfile5
#InputFileLocation = InputFileLocation1+InputFileLocation2+InputFileLocation3+InputFileLocation4+InputFileLocation5
#with open (file, 'r',encoding="utf-8") as file_input:
with open (InputFileLocation, 'r',encoding="utf-8") as file_input:
    Lines = file_input.readlines()
    #print (Lines)
cnt = 0
LineIndex = 0
PMID_LineIndex = []
for OneLine in Lines:
    LineIndex = LineIndex + 1
    if (OneLine.startswith('PMID-')):
        cnt = cnt +1
        PMID_LineIndex.append(LineIndex)
PMID_LineIndex.append(LineIndex)
#print(cnt)    
#print(PMID_LineIndex)

CancerType = "COVID-19"
Inputfile ="pubmed-COVID-19-set_10000.txt"
OutputFile = "pubmed-COVID-19-set_10000.xlsx"



InputDir = "../Src/"
OutputDir = "../Output/"

PUMID_List = []
TI_List = []
AB_List = []

#file = "../Src/pubmed-Top1-lungcancer-set.txt"
file = "../Src/pubmed-COVID-19-set_10000.txt"
with open (file, 'r',encoding="utf-8") as file_input:
    for i in range(0,cnt):
        StartLineIndex = PMID_LineIndex[i]
        EndLineIndex = PMID_LineIndex[i+1]
        OnePUMID = []
        OneTI_Info = []
        OneAB_Info = []
        for j in range(StartLineIndex,EndLineIndex):
            OneLine = file_input.readline()
            #OneLine = OneLine.strip()
            #print(OneLine)
            OnePUMID.append(OneLine)
        #print(len(OnePUMID))
        OnePubMed_LineCnt = len(OnePUMID)
        #print(OnePUMID)
        #print(OnePubMed_LineCnt)
        #OnePUMID_Info = []
        k = 0
        while (k < OnePubMed_LineCnt):
            if (OnePUMID[k].startswith('PMID-')):
                OnePUMID_Info = OnePUMID[k].split('- ')
                #print(str(k)+' PMID:'+OnePUMID_Info[1])
            if (OnePUMID[k].startswith('TI  -') or OnePUMID[k].startswith('Ti  -')):
                OneTI_Info = OnePUMID[k].split('- ')
                OneTI_Info[1] = OneTI_Info[1].strip()
                while (OnePUMID[k+1].startswith('  ')):
                    Temp = OnePUMID[k+1].split('   ')
                    #print(OnePUMID[k+1])
                    OneTI_Info[1] = OneTI_Info[1]+' '+Temp[1].strip()
                    k += 1
                #print(str(k)+' TI:'+OneTI_Info[1])
            if (OnePUMID[k].startswith('AB  -')):
                OneAB_Info = OnePUMID[k].split('- ')
                OneAB_Info[1] = OneAB_Info[1].strip()
                while (OnePUMID[k+1].startswith('  ')):
                    Temp = OnePUMID[k+1].split('   ')
                    #print(OnePUMID[k+1])
                    OneAB_Info[1] = OneAB_Info[1]+' '+Temp[1].strip()
                    k += 1
                #print(str(k)+' AB:'+OneTI_Info[1])
            k += 1

        PUMID_List.append(OnePUMID_Info[1].strip())
        TI_List.append(OneTI_Info[1].strip())
        if (len(OneAB_Info) == 2):
            AB_List.append(OneAB_Info[1].strip())
        else:
            AB_List.append('')
            #print(OnePUMID_Info[1]+"No Abstract")
        #AB_List.append(OneAB_Info[1].strip())
#PUMID_List
#TI_List
#AB_List
PubMed_Data = [[PUMID_List[i],TI_List[i],AB_List[i]] for i in range(cnt)]
#PubMed_Data

import pandas as pd

PubMed = pd.DataFrame(PubMed_Data, columns=['PUMID','TI','AB'])
#PubMed

PubMed['Type'] = CancerType

PubMed.to_excel("../Output/pubmed-COVID-19-set_10000.xlsx")

#### PubMed genome

#file = "../src/pubmed-Top1-lungcancer-set.txt"
InputDir = "../Src/"
#Inputfile1 ="pubmed-cancer-set_10000.txt"
#Inputfile2 ="pubmed-COVID-19-set_10000.txt"
Inputfile3 ="pubmed-genome-set_10000.txt"
#Inputfile4 ="pubmed-vaccine-set_10000.txt"
#Inputfile5 ="pubmed-virus-set_10000.txt"
#Inputfile = "pubmed-Top4_Prostateca-set.txt"

#InputFileLocation = InputDir+Inputfile1
#InputFileLocation = InputDir+Inputfile2
InputFileLocation = InputDir+Inputfile3
#InputFileLocation = InputDir+Inputfile4
#InputFileLocation = InputDir+Inputfile5
#InputFileLocation = InputFileLocation1+InputFileLocation2+InputFileLocation3+InputFileLocation4+InputFileLocation5
#with open (file, 'r',encoding="utf-8") as file_input:
with open (InputFileLocation, 'r',encoding="utf-8") as file_input:
    Lines = file_input.readlines()
    #print (Lines)
cnt = 0
LineIndex = 0
PMID_LineIndex = []
for OneLine in Lines:
    LineIndex = LineIndex + 1
    if (OneLine.startswith('PMID-')):
        cnt = cnt +1
        PMID_LineIndex.append(LineIndex)
PMID_LineIndex.append(LineIndex)
#print(cnt)    
#print(PMID_LineIndex)

CancerType = "Genome"
Inputfile ="pubmed-genome-set_10000.txt"
OutputFile = "pubmed-genome-set_10000.xlsx"



InputDir = "../Src/"
OutputDir = "../Output/"

PUMID_List = []
TI_List = []
AB_List = []

#file = "../Src/pubmed-Top1-lungcancer-set.txt"
file = "../Src/pubmed-genome-set_10000.txt"
with open (file, 'r',encoding="utf-8") as file_input:
    for i in range(0,cnt):
        StartLineIndex = PMID_LineIndex[i]
        EndLineIndex = PMID_LineIndex[i+1]
        OnePUMID = []
        OneTI_Info = []
        OneAB_Info = []
        for j in range(StartLineIndex,EndLineIndex):
            OneLine = file_input.readline()
            #OneLine = OneLine.strip()
            #print(OneLine)
            OnePUMID.append(OneLine)
        #print(len(OnePUMID))
        OnePubMed_LineCnt = len(OnePUMID)
        #print(OnePUMID)
        #print(OnePubMed_LineCnt)
        #OnePUMID_Info = []
        k = 0
        while (k < OnePubMed_LineCnt):
            if (OnePUMID[k].startswith('PMID-')):
                OnePUMID_Info = OnePUMID[k].split('- ')
                #print(str(k)+' PMID:'+OnePUMID_Info[1])
            if (OnePUMID[k].startswith('TI  -') or OnePUMID[k].startswith('Ti  -')):
                OneTI_Info = OnePUMID[k].split('- ')
                OneTI_Info[1] = OneTI_Info[1].strip()
                while (OnePUMID[k+1].startswith('  ')):
                    Temp = OnePUMID[k+1].split('   ')
                    #print(OnePUMID[k+1])
                    OneTI_Info[1] = OneTI_Info[1]+' '+Temp[1].strip()
                    k += 1
                #print(str(k)+' TI:'+OneTI_Info[1])
            if (OnePUMID[k].startswith('AB  -')):
                OneAB_Info = OnePUMID[k].split('- ')
                OneAB_Info[1] = OneAB_Info[1].strip()
                while (OnePUMID[k+1].startswith('  ')):
                    Temp = OnePUMID[k+1].split('   ')
                    #print(OnePUMID[k+1])
                    OneAB_Info[1] = OneAB_Info[1]+' '+Temp[1].strip()
                    k += 1
                #print(str(k)+' AB:'+OneTI_Info[1])
            k += 1

        PUMID_List.append(OnePUMID_Info[1].strip())
        TI_List.append(OneTI_Info[1].strip())
        if (len(OneAB_Info) == 2):
            AB_List.append(OneAB_Info[1].strip())
        else:
            AB_List.append('')
            #print(OnePUMID_Info[1]+"No Abstract")
        #AB_List.append(OneAB_Info[1].strip())
#PUMID_List
#TI_List
#AB_List
PubMed_Data = [[PUMID_List[i],TI_List[i],AB_List[i]] for i in range(cnt)]
#PubMed_Data

import pandas as pd

PubMed = pd.DataFrame(PubMed_Data, columns=['PUMID','TI','AB'])
#PubMed

PubMed['Type'] = CancerType

PubMed.to_excel("../Output/pubmed-genome-set_10000.xlsx")

#### PubMed vaccine

#file = "../src/pubmed-Top1-lungcancer-set.txt"
InputDir = "../Src/"
#Inputfile1 ="pubmed-cancer-set_10000.txt"
#Inputfile2 ="pubmed-COVID-19-set_10000.txt"
#Inputfile3 ="pubmed-genome-set_10000.txt"
Inputfile4 ="pubmed-vaccine-set_10000.txt"
#Inputfile5 ="pubmed-virus-set_10000.txt"
#Inputfile = "pubmed-Top4_Prostateca-set.txt"

#InputFileLocation = InputDir+Inputfile1
#InputFileLocation = InputDir+Inputfile2
#InputFileLocation = InputDir+Inputfile3
InputFileLocation = InputDir+Inputfile4
#InputFileLocation = InputDir+Inputfile5
#InputFileLocation = InputFileLocation1+InputFileLocation2+InputFileLocation3+InputFileLocation4+InputFileLocation5
#with open (file, 'r',encoding="utf-8") as file_input:
with open (InputFileLocation, 'r',encoding="utf-8") as file_input:
    Lines = file_input.readlines()
    #print (Lines)
cnt = 0
LineIndex = 0
PMID_LineIndex = []
for OneLine in Lines:
    LineIndex = LineIndex + 1
    if (OneLine.startswith('PMID-')):
        cnt = cnt +1
        PMID_LineIndex.append(LineIndex)
PMID_LineIndex.append(LineIndex)
#print(cnt)    
#print(PMID_LineIndex)

CancerType = "Vaccine"
Inputfile ="pubmed-vaccine-set_10000.txt"
OutputFile = "pubmed-vaccine-set_10000.xlsx"



InputDir = "../Src/"
OutputDir = "../Output/"

PUMID_List = []
TI_List = []
AB_List = []

#file = "../Src/pubmed-Top1-lungcancer-set.txt"
file = "../Src/pubmed-vaccine-set_10000.txt"
with open (file, 'r',encoding="utf-8") as file_input:
    for i in range(0,cnt):
        StartLineIndex = PMID_LineIndex[i]
        EndLineIndex = PMID_LineIndex[i+1]
        OnePUMID = []
        OneTI_Info = []
        OneAB_Info = []
        for j in range(StartLineIndex,EndLineIndex):
            OneLine = file_input.readline()
            #OneLine = OneLine.strip()
            #print(OneLine)
            OnePUMID.append(OneLine)
        #print(len(OnePUMID))
        OnePubMed_LineCnt = len(OnePUMID)
        #print(OnePUMID)
        #print(OnePubMed_LineCnt)
        #OnePUMID_Info = []
        k = 0
        while (k < OnePubMed_LineCnt):
            if (OnePUMID[k].startswith('PMID-')):
                OnePUMID_Info = OnePUMID[k].split('- ')
                #print(str(k)+' PMID:'+OnePUMID_Info[1])
            if (OnePUMID[k].startswith('TI  -') or OnePUMID[k].startswith('Ti  -')):
                OneTI_Info = OnePUMID[k].split('- ')
                OneTI_Info[1] = OneTI_Info[1].strip()
                while (OnePUMID[k+1].startswith('  ')):
                    Temp = OnePUMID[k+1].split('   ')
                    #print(OnePUMID[k+1])
                    OneTI_Info[1] = OneTI_Info[1]+' '+Temp[1].strip()
                    k += 1
                #print(str(k)+' TI:'+OneTI_Info[1])
            if (OnePUMID[k].startswith('AB  -')):
                OneAB_Info = OnePUMID[k].split('- ')
                OneAB_Info[1] = OneAB_Info[1].strip()
                while (OnePUMID[k+1].startswith('  ')):
                    Temp = OnePUMID[k+1].split('   ')
                    #print(OnePUMID[k+1])
                    OneAB_Info[1] = OneAB_Info[1]+' '+Temp[1].strip()
                    k += 1
                #print(str(k)+' AB:'+OneTI_Info[1])
            k += 1

        PUMID_List.append(OnePUMID_Info[1].strip())
        TI_List.append(OneTI_Info[1].strip())
        if (len(OneAB_Info) == 2):
            AB_List.append(OneAB_Info[1].strip())
        else:
            AB_List.append('')
            #print(OnePUMID_Info[1]+"No Abstract")
        #AB_List.append(OneAB_Info[1].strip())
#PUMID_List
#TI_List
#AB_List
PubMed_Data = [[PUMID_List[i],TI_List[i],AB_List[i]] for i in range(cnt)]
#PubMed_Data

import pandas as pd

PubMed = pd.DataFrame(PubMed_Data, columns=['PUMID','TI','AB'])
#PubMed

PubMed['Type'] = CancerType

PubMed.to_excel("../Output/pubmed-vaccine-set_10000.xlsx")

#### PubMed virus

#file = "../src/pubmed-Top1-lungcancer-set.txt"
InputDir = "../Src/"
#Inputfile1 ="pubmed-cancer-set_10000.txt"
#Inputfile2 ="pubmed-COVID-19-set_10000.txt"
#Inputfile3 ="pubmed-genome-set_10000.txt"
#Inputfile4 ="pubmed-vaccine-set_10000.txt"
Inputfile5 ="pubmed-virus-set_10000.txt"
#Inputfile = "pubmed-Top4_Prostateca-set.txt"

#InputFileLocation = InputDir+Inputfile1
#InputFileLocation = InputDir+Inputfile2
#InputFileLocation = InputDir+Inputfile3
#InputFileLocation = InputDir+Inputfile4
InputFileLocation = InputDir+Inputfile5
#InputFileLocation = InputFileLocation1+InputFileLocation2+InputFileLocation3+InputFileLocation4+InputFileLocation5
#with open (file, 'r',encoding="utf-8") as file_input:
with open (InputFileLocation, 'r',encoding="utf-8") as file_input:
    Lines = file_input.readlines()
    #print (Lines)
cnt = 0
LineIndex = 0
PMID_LineIndex = []
for OneLine in Lines:
    LineIndex = LineIndex + 1
    if (OneLine.startswith('PMID-')):
        cnt = cnt +1
        PMID_LineIndex.append(LineIndex)
PMID_LineIndex.append(LineIndex)
#print(cnt)    
#print(PMID_LineIndex)

CancerType = "Virus"
Inputfile ="pubmed-virus-set_10000.txt"
OutputFile = "pubmed-virus-set_10000.xlsx"



InputDir = "../Src/"
OutputDir = "../Output/"

PUMID_List = []
TI_List = []
AB_List = []

#file = "../Src/pubmed-Top1-lungcancer-set.txt"
file = "../Src/pubmed-virus-set_10000.txt"
with open (file, 'r',encoding="utf-8") as file_input:
    for i in range(0,cnt):
        StartLineIndex = PMID_LineIndex[i]
        EndLineIndex = PMID_LineIndex[i+1]
        OnePUMID = []
        OneTI_Info = []
        OneAB_Info = []
        for j in range(StartLineIndex,EndLineIndex):
            OneLine = file_input.readline()
            #OneLine = OneLine.strip()
            #print(OneLine)
            OnePUMID.append(OneLine)
        #print(len(OnePUMID))
        OnePubMed_LineCnt = len(OnePUMID)
        #print(OnePUMID)
        #print(OnePubMed_LineCnt)
        #OnePUMID_Info = []
        k = 0
        while (k < OnePubMed_LineCnt):
            if (OnePUMID[k].startswith('PMID-')):
                OnePUMID_Info = OnePUMID[k].split('- ')
                #print(str(k)+' PMID:'+OnePUMID_Info[1])
            if (OnePUMID[k].startswith('TI  -') or OnePUMID[k].startswith('Ti  -')):
                OneTI_Info = OnePUMID[k].split('- ')
                OneTI_Info[1] = OneTI_Info[1].strip()
                while (OnePUMID[k+1].startswith('  ')):
                    Temp = OnePUMID[k+1].split('   ')
                    #print(OnePUMID[k+1])
                    OneTI_Info[1] = OneTI_Info[1]+' '+Temp[1].strip()
                    k += 1
                #print(str(k)+' TI:'+OneTI_Info[1])
            if (OnePUMID[k].startswith('AB  -')):
                OneAB_Info = OnePUMID[k].split('- ')
                OneAB_Info[1] = OneAB_Info[1].strip()
                while (OnePUMID[k+1].startswith('  ')):
                    Temp = OnePUMID[k+1].split('   ')
                    #print(OnePUMID[k+1])
                    OneAB_Info[1] = OneAB_Info[1]+' '+Temp[1].strip()
                    k += 1
                #print(str(k)+' AB:'+OneTI_Info[1])
            k += 1

        PUMID_List.append(OnePUMID_Info[1].strip())
        TI_List.append(OneTI_Info[1].strip())
        if (len(OneAB_Info) == 2):
            AB_List.append(OneAB_Info[1].strip())
        else:
            AB_List.append('')
            #print(OnePUMID_Info[1]+"No Abstract")
        #AB_List.append(OneAB_Info[1].strip())
#PUMID_List
#TI_List
#AB_List
PubMed_Data = [[PUMID_List[i],TI_List[i],AB_List[i]] for i in range(cnt)]
#PubMed_Data

import pandas as pd

PubMed = pd.DataFrame(PubMed_Data, columns=['PUMID','TI','AB'])
#PubMed

PubMed['Type'] = CancerType

PubMed.to_excel("../Output/pubmed-virus-set_10000.xlsx")

#### PubMed genome

#file = "../src/pubmed-Top1-lungcancer-set.txt"
InputDir = "../Src/"
#Inputfile1 ="pubmed-cancer-set_10000.txt"
#Inputfile2 ="pubmed-COVID-19-set_10000.txt"
Inputfile3 ="pubmed-genome-set_10000.txt"
#Inputfile4 ="pubmed-vaccine-set_10000.txt"
#Inputfile5 ="pubmed-virus-set_10000.txt"
#Inputfile = "pubmed-Top4_Prostateca-set.txt"

#InputFileLocation = InputDir+Inputfile1
#InputFileLocation = InputDir+Inputfile2
InputFileLocation = InputDir+Inputfile3
#InputFileLocation = InputDir+Inputfile4
#InputFileLocation = InputDir+Inputfile5
#InputFileLocation = InputFileLocation1+InputFileLocation2+InputFileLocation3+InputFileLocation4+InputFileLocation5
#with open (file, 'r',encoding="utf-8") as file_input:
with open (InputFileLocation, 'r',encoding="utf-8") as file_input:
    Lines = file_input.readlines()
    #print (Lines)
cnt = 0
LineIndex = 0
PMID_LineIndex = []
for OneLine in Lines:
    LineIndex = LineIndex + 1
    if (OneLine.startswith('PMID-')):
        cnt = cnt +1
        PMID_LineIndex.append(LineIndex)
PMID_LineIndex.append(LineIndex)
#print(cnt)    
#print(PMID_LineIndex)

CancerType = "Genome"
Inputfile ="pubmed-genome-set_10000.txt"
OutputFile = "pubmed-genome-set_10000.xlsx"



InputDir = "../Src/"
OutputDir = "../Output/"

PUMID_List = []
TI_List = []
AB_List = []

#file = "../Src/pubmed-Top1-lungcancer-set.txt"
file = "../Src/pubmed-genome-set_10000.txt"
with open (file, 'r',encoding="utf-8") as file_input:
    for i in range(0,cnt):
        StartLineIndex = PMID_LineIndex[i]
        EndLineIndex = PMID_LineIndex[i+1]
        OnePUMID = []
        OneTI_Info = []
        OneAB_Info = []
        for j in range(StartLineIndex,EndLineIndex):
            OneLine = file_input.readline()
            #OneLine = OneLine.strip()
            #print(OneLine)
            OnePUMID.append(OneLine)
        #print(len(OnePUMID))
        OnePubMed_LineCnt = len(OnePUMID)
        #print(OnePUMID)
        #print(OnePubMed_LineCnt)
        #OnePUMID_Info = []
        k = 0
        while (k < OnePubMed_LineCnt):
            if (OnePUMID[k].startswith('PMID-')):
                OnePUMID_Info = OnePUMID[k].split('- ')
                #print(str(k)+' PMID:'+OnePUMID_Info[1])
            if (OnePUMID[k].startswith('TI  -') or OnePUMID[k].startswith('Ti  -')):
                OneTI_Info = OnePUMID[k].split('- ')
                OneTI_Info[1] = OneTI_Info[1].strip()
                while (OnePUMID[k+1].startswith('  ')):
                    Temp = OnePUMID[k+1].split('   ')
                    #print(OnePUMID[k+1])
                    OneTI_Info[1] = OneTI_Info[1]+' '+Temp[1].strip()
                    k += 1
                #print(str(k)+' TI:'+OneTI_Info[1])
            if (OnePUMID[k].startswith('AB  -')):
                OneAB_Info = OnePUMID[k].split('- ')
                OneAB_Info[1] = OneAB_Info[1].strip()
                while (OnePUMID[k+1].startswith('  ')):
                    Temp = OnePUMID[k+1].split('   ')
                    #print(OnePUMID[k+1])
                    OneAB_Info[1] = OneAB_Info[1]+' '+Temp[1].strip()
                    k += 1
                #print(str(k)+' AB:'+OneTI_Info[1])
            k += 1

        PUMID_List.append(OnePUMID_Info[1].strip())
        TI_List.append(OneTI_Info[1].strip())
        if (len(OneAB_Info) == 2):
            AB_List.append(OneAB_Info[1].strip())
        else:
            AB_List.append('')
            #print(OnePUMID_Info[1]+"No Abstract")
        #AB_List.append(OneAB_Info[1].strip())
#PUMID_List
#TI_List
#AB_List
PubMed_Data = [[PUMID_List[i],TI_List[i],AB_List[i]] for i in range(cnt)]
#PubMed_Data

import pandas as pd

PubMed = pd.DataFrame(PubMed_Data, columns=['PUMID','TI','AB'])
#PubMed

PubMed['Type'] = CancerType

PubMed.to_excel("../Output/pubmed-genome-set-10000.xlsx")
