==> 1_Merge 5 articles.ipynb <==
TypeArticles = [
"../Output/pubmed-cancer-set_10000.xlsx",
"../Output/pubmed-COVID-19-set_10000.xlsx",
"../Output/pubmed-genome-set_10000.xlsx",
"../Output/pubmed-vaccine-set_10000.xlsx",
"../Output/pubmed-virus-set_10000.xlsx",
]
TypeArticles

len(TypeArticles)

import pandas as pd

PubMed_Articles = pd.read_excel(TypeArticles[4])
PubMed_Articles

import pandas as pd

PubMed_Articles = pd.read_excel(TypeArticles[0])
#PubMed_Articles
for i in range (1,len(TypeArticles)):
    print(TypeArticles[i])
    OnePubMed_Articles = pd.read_excel(TypeArticles[i])
    PubMed_Articles = PubMed_Articles.append(OnePubMed_Articles)

PubMed_Articles

PubMed_Articles.shape

OutputFileName = "PubMedParsedSrc_"+str(len(TypeArticles))+"_classes"
OutputFileName

OutputLoction = "../Src_Parsed/"+OutputFileName+".xlsx"
OutputLoction

PubMed_Articles.to_excel(OutputLoction)
