==> 2_Preprocessing.ipynb <==
import pandas as pd
import numpy as np
import re
import nltk
import matplotlib.pyplot as plt

pd.options.display.max_colwidth = 200

PubMed_Articles_Loction = "../Src_Parsed/PubMedParsedSrc_5_classes.xlsx"
PubMed_Articles = pd.read_excel(PubMed_Articles_Loction)

PubMed_Articles.head()

PubMed_Articles['Content'] = PubMed_Articles['TI']+ " "+PubMed_Articles['AB']
PubMed_Articles

PubMed_Articles = PubMed_Articles.reindex(columns = ['PUMID','TI','AB','Content','Type'])
PubMed_Articles

labels = PubMed_Articles['Type'].tolist()
#print(labels)
ClassIDMap = {'Cancer': 1, 'COVID-19':2, 'Genome':3, 'Vaccine': 4, 'Virus': 5 }
ClassIDMap

corpus, target_labels, target_names = (PubMed_Articles['Content'], [ClassIDMap[label] for label in PubMed_Articles['Type']], PubMed_Articles['Type'])
data_df = pd.DataFrame({'Article': corpus, 'Target Label': target_labels, 'Target Name': target_names})
print(data_df.shape)
data_df.tail(25)

data_df.info()

# just to keep negation if any in bi-grams
wpt = nltk.WordPunctTokenizer()
stop_words = nltk.corpus.stopwords.words('english')

def normalize_document(doc):
    # lower case and remove special characters\whitespaces
    doc = re.sub(r'[^a-zA-Z\s]', '', doc, re.I|re.A)
    doc = doc.lower()
    doc = doc.strip()
    # tokenize document
    tokens = wpt.tokenize(doc)
    # filter stopwords out of document
    filtered_tokens = [token for token in tokens if token not in stop_words]
    # re-create document from filtered tokens
    doc = ' '.join(filtered_tokens)
    return doc

normalize_corpus = np.vectorize(normalize_document)

data_df['Article']=data_df['Article'].apply(str)
data_df['normalize_article'] = normalize_corpus(data_df['Article'])

data_df = data_df[['Article', 'normalize_article', 'Target Label', 'Target Name']]
data_df.tail(40)

import nltk

w_tokenizer = nltk.tokenize.WhitespaceTokenizer()
lemmatizer = nltk.stem.WordNetLemmatizer()

def lemmatize_text(text):
    return [lemmatizer.lemmatize(w) for w in w_tokenizer.tokenize(text)]

data_df['lemmatized'] = data_df.normalize_article.apply(lemmatize_text)

data_df = data_df[['Article', 'lemmatized', 'Target Label', 'Target Name']]
data_df.tail(40)

data_df.info()

data_df = data_df.replace(r'^(\s?)+$', np.nan, regex=True)
data_df.info()

OutputFileName = "PubMedArticle(normalize+lemmatize)"
OutputFileName

OutputLoction = "../Output/"+OutputFileName+".csv"
OutputLoction

data_df.to_csv(OutputLoction)
