==> 3_classification_TF.ipynb <==
import pandas as pd
#data_df = pd.read_csv('clean_newsgroups.csv')
data_df = pd.read_csv('../Output/PubMedCleanArticle(normalize+lemmetize).csv', encoding='utf-8')

data_df.head(10)

import numpy as np
np.array(data_df['Target Label'])

np.array(data_df['Target Name'])

import numpy as np
from sklearn.model_selection import train_test_split

train_corpus, test_corpus, train_label_nums, test_label_nums, train_label_names, test_label_names =\
                                 train_test_split(np.array(data_df['lemmatized']), np.array(data_df['Target Label']),
                                                       np.array(data_df['Target Name']), test_size=0.33, random_state=42)

train_corpus.shape, test_corpus.shape

train_label_nums, test_label_nums

train_label_names, test_label_names

from sklearn.feature_extraction.text import CountVectorizer
# get bag of words features in sparse format
cv = CountVectorizer(min_df=0., max_df=1.)
cv_matrix = cv.fit_transform(data_df['lemmatized'])
cv_matrix

from collections import Counter

trd = dict(Counter(train_label_names))
tsd = dict(Counter(test_label_names))

(pd.DataFrame([[key, trd[key], tsd[key]] for key in trd], 
             columns=['Target Label', 'Train Count', 'Test Count'])
.sort_values(by=['Train Count', 'Test Count'],
             ascending=False))

import nltk
(pd.Series(nltk.ngrams(data_df['lemmatized'], 2)).value_counts())[:3]

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.model_selection import cross_val_score

# build BOW features on train articles
cv = CountVectorizer(binary=False, min_df=0.00002, max_df=0.6)
cv_train_features = cv.fit_transform(train_corpus)

print(cv_train_features.shape)

print(cv_train_features)

# transform test articles into features
cv_test_features = cv.transform(test_corpus.astype('U'))
print(cv_test_features.shape)
print(cv_train_features)

print('BOW model:> Train features shape:', cv_train_features.shape, ' Test features shape:', cv_test_features.shape)

from sklearn.linear_model import LogisticRegression

lr = LogisticRegression(penalty='l2', max_iter=1000, C=0.1, random_state=42)
lr.fit(cv_train_features, train_label_names)
lr_bow_cv_scores = cross_val_score(lr, cv_train_features, train_label_names, cv=5)
lr_bow_cv_mean_score = np.mean(lr_bow_cv_scores)
print('CV Accuracy (5-fold):', lr_bow_cv_scores)
print('Mean CV Accuracy:', lr_bow_cv_mean_score)
lr_bow_test_score = lr.score(cv_test_features, test_label_names)
print('Test Accuracy:', lr_bow_test_score)

from sklearn.svm import LinearSVC

svm = LinearSVC(penalty='l2', max_iter=100, C=0.001, random_state=42)
svm.fit(cv_train_features, train_label_names)
svm_bow_cv_scores = cross_val_score(svm, cv_train_features, train_label_names, cv=5)
svm_bow_cv_mean_score = np.mean(svm_bow_cv_scores)
print('CV Accuracy (5-fold):', svm_bow_cv_scores)
print('Mean CV Accuracy:', svm_bow_cv_mean_score)
svm_bow_test_score = svm.score(cv_test_features, test_label_names)
print('Test Accuracy:', svm_bow_test_score)

pd.DataFrame([['Logistic Regression', lr_bow_cv_mean_score, lr_bow_test_score],
              ['Linear SVM', svm_bow_cv_mean_score, svm_bow_test_score]],
              columns=['Model',  'CV Score (TF)', 'Test Score (TF)'],
             )
