==> 5_Classification_TFIDF_Parameters.ipynb <==
import pandas as pd
#data_df = pd.read_csv('clean_newsgroups.csv')
data_df = pd.read_csv('../Output/PubMedArticle(normalize+lemmatize).csv', encoding='utf-8')
data_df.head(10)

data_labels_map = dict(enumerate(pd.unique(data_df['Target Name'])))
data_labels_map

import numpy as np
from sklearn.model_selection import train_test_split

train_corpus, test_corpus, train_label_nums, test_label_nums, train_label_names, test_label_names =\
                                 train_test_split(np.array(data_df['lemmatized']), np.array(data_df['Target Label']),
                                                       np.array(data_df['Target Name']), test_size=0.33, random_state=42)

train_corpus.shape, test_corpus.shape

from sklearn.feature_extraction.text import TfidfVectorizer

# build BOW features on train articles
tv = TfidfVectorizer(use_idf=True, min_df=0.0, max_df=1.0)
tv_train_features = tv.fit_transform(train_corpus.astype('U'))
# transform test articles into features
tv_test_features = tv.transform(test_corpus.astype('U'))
print('TFIDF model:> Train features shape:', tv_train_features.shape, ' Test features shape:', tv_test_features.shape)

from sklearn.pipeline import Pipeline
from sklearn.svm import LinearSVC ,SVC
from sklearn.model_selection import GridSearchCV


svm_pipeline = Pipeline([('tfidf', TfidfVectorizer()),
                        ('svm', LinearSVC(random_state=42))])

param_grid = {'tfidf__ngram_range': [(1, 1), (1, 1)],
              'svm__C': [0.01, 0.1, 1, 5]
}

gs_svm = GridSearchCV(svm_pipeline, param_grid, cv=5, verbose=2)
gs_svm = gs_svm.fit(train_corpus.astype('U'), train_label_names)

gs_svm.best_estimator_.get_params()

best_svm_test_score = gs_svm.score(test_corpus.astype('U'), test_label_names)
print('Test Accuracy :', best_svm_test_score)

from sklearn.linear_model import LogisticRegression, SGDClassifier
from sklearn.svm import LinearSVC
from sklearn.pipeline import Pipeline
from sklearn.model_selection import GridSearchCV

lr_pipeline = Pipeline([('tfidf', TfidfVectorizer()),
                        ('lr', LogisticRegression(penalty='l2', max_iter=1000, random_state=42))
                       ])

param_grid = {'tfidf__ngram_range': [(1, 1), (1, 1)],
              'lr__C': [1, 5, 10]
}

gs_lr = GridSearchCV(lr_pipeline, param_grid, cv=5, verbose=2)
gs_lr = gs_lr.fit(train_corpus.astype('U'), train_label_names)

gs_lr.best_estimator_.get_params()

best_lr_test_score = gs_lr.score(test_corpus.astype('U'), test_label_names)
print('Test Accuracy :', best_lr_test_score)

import model_evaluation_utils as meu


svm_predictions = gs_svm.predict(test_corpus.astype('U'))
unique_classes = list(set(test_label_names))
meu.get_metrics(true_labels=test_label_names, predicted_labels=svm_predictions)

meu.display_classification_report(true_labels=test_label_names, 
                                  predicted_labels=svm_predictions, classes=unique_classes)

lr_predictions = gs_lr.predict(test_corpus.astype('U'))
unique_classes = list(set(test_label_names))
meu.get_metrics(true_labels=test_label_names, predicted_labels=lr_predictions)

meu.display_classification_report(true_labels=test_label_names, 
                                  predicted_labels=lr_predictions, classes=unique_classes)

label_data_map = {v:k for k, v in data_labels_map.items()}
label_map_df = pd.DataFrame(list(label_data_map.items()), columns=['Label Name', 'Label Number'])
label_map_df

train_idx, test_idx = train_test_split(np.array(range(len(data_df['lemmatized']))), test_size=0.33, random_state=42)
test_idx

predict_probas = gs_lr.predict_proba(test_corpus.astype('U')).max(axis=1)
test_df = data_df.loc[test_idx]
test_df['Predicted Name'] = lr_predictions
test_df['Predicted Confidence'] = predict_probas
test_df.head()

pd.set_option('display.max_colwidth', 200)
res_df = (test_df[(test_df['Target Name'] == 'Cancer') & (test_df['Predicted Name'] == 'Vaccine')]
       .sort_values(by=['Predicted Confidence'], ascending=False).head(5))
res_df

pd.set_option('display.max_colwidth', 200)
res_df = (test_df[(test_df['Target Name'] == 'Virus') & (test_df['Predicted Name'] == 'Covid-19')]
       .sort_values(by=['Predicted Confidence'], ascending=False).head(5))
res_df

pd.set_option('display.max_colwidth', 200)
res_df = (test_df[(test_df['Target Name'] == 'Virus') & (test_df['Predicted Name'] == 'Genome')]
       .sort_values(by=['Predicted Confidence'], ascending=False).head(5))
res_df
